requirements.txt
gpsd-py3==0.3.0
